Finish writing unit tests for all API operations

+++ recipe/serializers.py
@@ -1,7 +1,6 @@
 from django.db.models.query import QuerySet
 from rest_framework import serializers
 from core.models import *
-from rest_framework.fields import SerializerMethodField
 
 
 class IngredientSerializer(serializers.ModelSerializer):
@@ -14,7 +13,7 @@ class IngredientSerializer(serializers.ModelSerializer):
 
 class RecipeSerializer(serializers.ModelSerializer):
     """Serialize a recipe"""
-    ingredients = IngredientSerializer(source='ingredient_set', many=True)
+    ingredients = IngredientSerializer(many=True)
     
     class Meta:
         model = Recipe
@@ -22,20 +21,22 @@ class RecipeSerializer(serializers.ModelSerializer):
         read_only_fields = ('id',)
 
     def create(self, data):
-        ingredients_data = data.pop('ingredient_set')
         recipe = Recipe.objects.create(**data)
-        for i in ingredients_data:
-            Ingredient.objects.create(**i, recipe=recipe)
+        # if 'ingredients' in data:
+        #     ingredients_data = data.pop('ingredients')
+        #     for i in ingredients_data:
+        #         Ingredient.objects.create(**i, recipe=recipe)
         return recipe
 
     def update(self, instance, data):
         instance.name = data.get('name', instance.name)
         instance.description = data.get('description', instance.description)
-        ingredients_data = data.pop('ingredient_set')
-        if ingredients_data:
-            Ingredient.objects.filter(recipe=instance).delete()
-            for i in ingredients_data:
-                Ingredient.objects.create(**i, recipe=instance)
-        instance.save()
+        if 'ingredients' in data:
+            ingredients_data = data.pop('ingredients')
+            if ingredients_data:
+                Ingredient.objects.filter(recipe=instance).delete()
+                for i in ingredients_data:
+                    Ingredient.objects.create(**i, recipe=instance)
+            instance.save()
         return instance
